backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
aiofiles==24.1.0
python-multipart==0.0.17
pillow==11.0.0
numpy==1.26.4
opencv-python==4.10.0.84
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.0
celery==5.4.0
ollama==0.4.2
qwen-vl-utils==0.0.8
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
--extra-index-url https://download.pytorch.org/whl/cu121
